spike: S12.25 switch category axis to macros (BASE+n, embedded cast)
Categories become uint16_t macros carrying their own cast, so emit sites
drop the (uint16_t) clutter and BASE+n arithmetic is restored. cppcheck-misra
clean on the UdpSender TUs with zero new suppressions (vs ~2 D.009 lines per
header for the enum variant). Detail stays a per-class enum with a visible
(int32_t) cast — it is polymorphic (enum today, native errno/X509 later).

+++ Core/Source/SolidSyslogUdpSender.c
@@ -76,7 +76,7 @@ static bool UdpSender_Send(struct SolidSyslogSender* base, const void* buffer, s
         SolidSyslog_Error(
             SOLIDSYSLOG_SEVERITY_ERROR,
             &UdpSenderErrorSource,
-            (uint16_t) SOLIDSYSLOG_CAT_SENDER_SEND_NULL_BUFFER,
+            SOLIDSYSLOG_CAT_SENDER_SEND_NULL_BUFFER,
             (int32_t) UDPSENDER_ERROR_SEND_NULL_BUFFER
         );
     }
